that's meant to be a byte not an int

+++ src/Maths.java
@@ -19,7 +19,7 @@ public class Maths {
             for(d = 0; d < n; d++)
                 matrix[c][d] = in.nextByte();
 
-        int transpose[][] = new int[n][m];
+        byte transpose[][] = new byte[n][m];
 
         for(c = 0; c < m; c++){
             for(d = 0; d < n; d++)
